fix(internal/repository/user_repository): refactor FindAllUser to return UserResponse
This commit refactors the FindAllUser method in the user_repository to return a slice of UserResponse instead of a slice of User. This change is necessary because the UserResponse type contains additional information that is needed by the frontend.

The following changes were made to implement this refactor:

- The FindAllUser method signature was changed to return a slice of UserResponse.
- The FindAllUser method body was updated to convert the slice of User to a slice of UserResponse.
- The UserResponse type was updated to include the additional information that is needed by the frontend.

This refactor does not affect the functionality of the FindAllUser method. However, it does improve the usability of the method by providing the frontend with the additional information that it needs.