refactor(user): change Find* method return type
This commit refactors the Find* methods in the user repository and usecase to return entity.UserResponse instead of entity.User.

This change was made to simplify the API and make it more consistent with other parts of the codebase. Previously, the Find* methods returned an entity.User, which contained all of the user's data. However, in most cases, only a subset of this data is needed. By returning an entity.UserResponse, we can reduce the amount of data that is returned and improve the performance of the API.